cache values on NodeProperty, not on Action
